ingest/requirements.txt
pandas==2.2.2
pyarrow==16.1.0
requests==2.32.3
python-dateutil==2.9.0